refactor(cli): extract shared captureStream test helper
Twelve test files across @baerly/cli and @gusto/create-baerly-storage
each carried a byte-identical captureStream helper (~16 lines) that
swaps process.stdout/stderr write for a buffer during a call. Hoist it
into a single test-only module at
packages/cli/src/_internal/testing.ts, exposed via the
"./_internal/testing" subpath (mirroring @baerly/server/_internal/testing;
intentionally not in publishConfig).

cli test files import it relatively; the two create-baerly-storage tests
import @baerly/cli/_internal/testing over the existing @baerly/cli devDep
(an edge the package-layer lint already permits). Net -168 lines of
duplication with no behavior change.
