+++ packages/cli/src/admin/fsck.test.ts
@@ -27,6 +27,7 @@ import { LocalFsStorage } from "@baerly/dev";
 import { allIndexKeysFor } from "@baerly/server";
 import { Writer } from "@baerly/server/_internal/testing";
 import { runFsck } from "./fsck.ts";
+import { captureStream } from "../_internal/testing.ts";
 
 const APP = "app";
 const TENANT = "tenant";
@@ -60,23 +61,6 @@ const seedRows = async (storage: Storage, count: number): Promise<void> => {
       body: { _id: `t-${i}`, status: i % 2 === 0 ? "open" : "closed" },
     });
   }
-};
-
-const captureStream = (
-  stream: NodeJS.WriteStream,
-): { restore: () => void; readonly captured: string[] } => {
-  const captured: string[] = [];
-  const original = stream.write.bind(stream);
-  stream.write = ((chunk: unknown): boolean => {
-    captured.push(typeof chunk === "string" ? chunk : String(chunk));
-    return true;
-  }) as typeof stream.write;
-  return {
-    captured,
-    restore: () => {
-      stream.write = original;
-    },
-  };
 };
 
 describe("baerly admin fsck — consistency check and corruption findings", () => {
