+++ packages/cli/src/admin/rebuild-index.test.ts
@@ -24,23 +24,7 @@ import { LocalFsStorage } from "@baerly/dev";
 import { allIndexKeysFor } from "@baerly/server";
 import { Writer } from "@baerly/server/_internal/testing";
 import { runRebuildIndex } from "./rebuild-index.ts";
-
-const captureStream = (
-  stream: NodeJS.WriteStream,
-): { restore: () => void; readonly captured: string[] } => {
-  const captured: string[] = [];
-  const original = stream.write.bind(stream);
-  stream.write = ((chunk: unknown): boolean => {
-    captured.push(typeof chunk === "string" ? chunk : String(chunk));
-    return true;
-  }) as typeof stream.write;
-  return {
-    captured,
-    restore: () => {
-      stream.write = original;
-    },
-  };
-};
+import { captureStream } from "../_internal/testing.ts";
 
 const APP = "app";
 const TENANT = "tenant";
